refactor: move functions for overlay to own module manage_overlay.js
